# fix(npm): Comprehensive solution for Issue #155 - ruv-swarm package issues