… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/Cargo.toml
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/README.md
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/benches/coordinator_benchmarks.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/README.md
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/basic_coordination.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/basic_swarm.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/browser/forecasting.html
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/browser/index.html
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/cognitive_diversity.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/daa-service-demo.js
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/fibonacci_demo.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/forecasting_demo.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/geometric_langlands_demo.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/github_integration.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/neural_swarm_integration.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/neural_training.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/performance_monitoring.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/performance_monitoring_demo.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/wasm_swarm_demo.js
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/web_demo/demo.js
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/examples/web_demo/index.html
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/src/enhanced_queen_coordinator.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/src/lib.rs
… ruv-swarm/crates/ruv-swarm-enhanced-coordinator/tests/enhanced_coordinator_tests.rs
… ruv-swarm/npm/CRITICAL_REVIEW_ISSUE_155.md
… ruv-swarm/npm/MCP_SERVER_FIX_SUMMARY.md
… ruv-swarm/npm/PERFORMANCE_OPTIMIZATION_REPORT.md
… ruv-swarm/npm/VALIDATION_REPORT_163.md
… ruv-swarm/npm/data/ruv-swarm.db-shm
… ruv-swarm/npm/docs/ES_MODULE_MIGRATION_REPORT.md